Component Work (User and Course)
Worked on UserSignIn and UserSignUp Components in detail.  Outlined Update Course Component, but having issues, so some code is commented out.  Created a file for HelperFunctions, but not currently using it.  Updated notes in Courses Component.  Updated CourseDetail Component to use async/await in the componentDidMount function, and outlined future plans